embedded/requirements.txt
Adafruit-Blinka==8.31.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-ccs811==1.3.15
adafruit-circuitpython-gps==3.10.12
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==2.0.4
adafruit-circuitpython-typing==1.9.6
Adafruit-PlatformDetect==3.59.0
Adafruit-PureIO==1.1.11
board==1.0
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
cycler==0.10.0
distro==1.5.0
gpiozero==1.6.2
gps==3.22
idna==2.10
kiwisolver==1.3.1
logging==0.4.9.6
matplotlib==3.3.4
numpy==1.19.5
paho-mqtt==1.6.1
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pycairo==1.16.2
pyftdi==0.55.0
PyGObject==3.38.0
pyparsing==2.4.7
pyserial==3.5
python-apt==2.2.1
python-dateutil==2.8.1
python-prctl==1.7
pyusb==1.2.1
requests==2.25.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
smbus2==0.4.3
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
typing-extensions==4.9.0
urllib3==1.26.5
v4l2-python3==0.3.2
